requirements.txt
modbus-tk
paho-mqtt
apscheduler